adapt hip two different bfloat16

+++ common/unified/components/fill_array_kernels.cpp
@@ -43,7 +43,8 @@ void fill_seq_array(std::shared_ptr<const DefaultExecutor> exec,
         [] GKO_KERNEL(auto idx, auto array) {
             if constexpr (std::is_same_v<remove_complex<ValueType>, float16>) {
                 // __half can not be from int64_t
-                array[idx] = static_cast<long long>(idx);
+                // __hip_bfloat16 can not be from long long
+                array[idx] = static_cast<float>(idx);
             } else {
                 array[idx] = idx;
             }
